[#648] Refactor remaining view templates
- Updated `template_name` for multiple views from `:vue_point` to `:index`.
- Moved `Burn` view class from `views.rb` to `deprecated.rb` and updated its `template_name`.
- Renamed `Meta::Robot` view class to `RobotsTxt` and updated its template details.
- Modified `not_found_response` to use `Error` view instead of `NotFound`.
- Adjusted `robots_txt` endpoint to use the new `RobotsTxt` view class.
